requirements.txt
streamlit
pandas
numpy
scikit-learn
mysql-connector-python
PyPDF2
python-docx
python-dotenv
